+++ src/Products/PloneMeeting/ToolPloneMeeting.py
@@ -24,7 +24,6 @@ from collective.iconifiedcategory.utils import get_config_root
 from collective.iconifiedcategory.utils import update_all_categorized_elements
 from datetime import datetime
 from DateTime import DateTime
-from ftw.labels.interfaces import ILabeling
 from ftw.labels.labeling import ANNOTATION_KEY as FTW_LABELS_ANNOTATION_KEY
 from imio.actionspanel.utils import unrestrictedRemoveGivenObject
 from imio.helpers.cache import cleanForeverCache
@@ -41,7 +40,6 @@ from imio.helpers.security import check_zope_admin
 from imio.helpers.security import fplog
 from imio.history.utils import add_event_to_wf_history
 from imio.migrator.utils import end_time
-from imio.prettylink.interfaces import IPrettyLink
 from OFS import CopySupport
 from persistent.mapping import PersistentMapping
 from plone import api
@@ -94,7 +92,6 @@ from Products.PloneMeeting.utils import notifyModifiedAndReindex
 from Products.PloneMeeting.utils import org_id_to_uid
 from Products.PloneMeeting.utils import reindex_object
 from Products.PloneMeeting.utils import workday
-from Products.ZCatalog.Catalog import AbstractCatalogBrain
 from ZODB.POSException import ConflictError
 from zope.annotation.interfaces import IAnnotations
 from zope.i18n import translate
@@ -877,26 +874,6 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
         else:
             return True
 
-    def getUserName_cachekey(method, self, userId, withUserId=False):
-        '''cachekey method for self.getUserName.'''
-        return userId, withUserId
-
-    security.declarePublic('getUserName')
-
-    # @ram.cache(getUserName_cachekey)
-    def getUserName(self, userId, withUserId=False):
-        '''Returns the full name of user having id p_userId.
-           Performance test does not show that ram.cache is necessary.'''
-        res = get_user_fullname(userId)
-        # fullname of a Zope user (admin) is returned as unicode
-        # and fullname of a Plone user is returned as utf-8...
-        # always return as utf-8!
-        if isinstance(res, unicode):
-            res = res.encode('utf-8')
-        if withUserId:
-            res = res + " ({0})".format(userId)
-        return res
-
     security.declarePrivate('listWeekDays')
 
     def listWeekDays(self):
@@ -1509,31 +1486,6 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
                 history.append(event)
         obj.workflow_history[workflow_name] = tuple(history)
 
-    def getAvailableMailingLists(self, obj, pod_template):
-        '''Gets the names of the (currently active) mailing lists defined for
-           this template.'''
-        res = []
-        mailing_lists = pod_template.mailing_lists and pod_template.mailing_lists.strip()
-        if not mailing_lists:
-            return res
-        try:
-            extra_expr_ctx = _base_extra_expr_ctx(obj)
-            extra_expr_ctx.update({'obj': obj, })
-            for line in mailing_lists.split('\n'):
-                name, expression, userIds = line.split(';')
-                if not expression or _evaluateExpression(obj,
-                                                         expression,
-                                                         roles_bypassing_expression=[],
-                                                         extra_expr_ctx=extra_expr_ctx,
-                                                         raise_on_error=True):
-                    res.append(name.strip())
-        except Exception, exc:
-            res.append(translate('Mailing lists are not correctly defined, original error is \"${error}\"',
-                                 domain='PloneMeeting',
-                                 mapping={'error': str(exc)},
-                                 context=self.REQUEST))
-        return res
-
     def showHolidaysWarning(self, cfg):
         """Condition for showing the 'holidays_waring_message'."""
         if cfg is not None and cfg.__class__.__name__ == "MeetingConfig":
@@ -1560,31 +1512,6 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
     def get_extra_adviser_infos(self):
         '''See doc in interfaces.py.'''
         return {}
-
-    def getAdvicePortalTypeIds_cachekey(method, self):
-        '''cachekey method for self.getAdvicePortalTypes.'''
-        return True
-
-    security.declarePublic('getAdvicePortalTypeIds')
-
-    @ram.cache(getAdvicePortalTypeIds_cachekey)
-    def getAdvicePortalTypeIds(self):
-        """We may have several 'meetingadvice' portal_types,
-           return it as ids."""
-        return self.getAdvicePortalTypes(as_ids=True)
-
-    security.declarePublic('getAdvicePortalTypes')
-
-    def getAdvicePortalTypes(self, as_ids=False):
-        """We may have several 'meetingadvice' portal_types."""
-        typesTool = api.portal.get_tool('portal_types')
-        res = []
-        for portal_type in typesTool.listTypeInfo():
-            if portal_type.id.startswith('meetingadvice'):
-                res.append(portal_type)
-        if as_ids:
-            res = [p.id for p in res]
-        return res
 
     def getGroupedConfigs_cachekey(method, self, config_group=None, check_access=True, as_items=False):
         '''cachekey method for self.getGroupedConfigs.'''
@@ -1633,22 +1560,6 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
         res = True
         if not check_zope_admin():
             res = False
-        return res
-
-    def get_labels(self, obj, include_personal_labels=True):
-        """Return active labels for p_obj.
-           p_include_personal_labels may be:
-           - True: returns every labels, personal or not;
-           - False: personal labels not returned;
-           - "only": only personal labels returned."""
-        res = {}
-        labeling = ILabeling(obj)
-        labels = labeling.active_labels()
-        for label in labels:
-            if (include_personal_labels == "only" and not label['by_user']) or \
-               (include_personal_labels is False and label['by_user']):
-                continue
-            res[label['label_id']] = label['title']
         return res