+++ src/Products/PloneMeeting/content/advice.py
@@ -18,6 +18,7 @@ from Products.PloneMeeting.config import PMMessageFactory as _
 from Products.PloneMeeting.interfaces import IDXMeetingContent
 from Products.PloneMeeting.utils import findMeetingAdvicePortalType
 from Products.PloneMeeting.utils import get_event_field_data
+from Products.PloneMeeting.utils import getAdvicePortalTypeIds
 from Products.PloneMeeting.utils import getWorkflowAdapter
 from Products.PloneMeeting.utils import historize_object_data
 from Products.PloneMeeting.utils import isModifiedSinceLastVersion
@@ -286,8 +287,7 @@ class AdviceGroupVocabulary(object):
     def __call__(self, context):
         """"""
         terms = []
-        tool = api.portal.get_tool('portal_plonemeeting')
-        advicePortalTypeIds = tool.getAdvicePortalTypeIds()
+        advicePortalTypeIds = getAdvicePortalTypeIds()
 
         # take into account groups for wich user can add an advice
         # while adding an advice, the context is his parent, aka a MeetingItem
@@ -327,7 +327,6 @@ class AdviceTypeVocabulary(object):
         terms = []
         tool = api.portal.get_tool('portal_plonemeeting')
         cfg = tool.getMeetingConfig(context)
-        advicePortalTypeIds = tool.getAdvicePortalTypeIds()
 
         # manage when portal_type accessed from the Dexterity types configuration
         if cfg:
@@ -342,7 +341,7 @@ class AdviceTypeVocabulary(object):
 
             # make sure if an adviceType was used for context and it is no more available, it
             # appears in the vocabulary and is so useable...
-            if context.portal_type in advicePortalTypeIds and \
+            if context.portal_type in getAdvicePortalTypeIds() and \
                context.advice_type not in usedAdviceTypes:
                 usedAdviceTypes.append(context.advice_type)
             for advice_id, advice_title in cfg.listAdviceTypes(include_asked_again=True).items():